tools/gw/src/requirements.txt
google-api-python-client
google-auth-oauthlib
click
